+++ memori/__main__.py
@@ -18,6 +18,7 @@ from memori._config import Config
 from memori._setup import Manager as SetupManager
 from memori.api._quota import Manager as ApiQuotaManager
 from memori.api._sign_up import Manager as ApiSignUpManager
+from memori.provisioning._manager import Manager as ProvisioningManager
 from memori.storage.cockroachdb._cluster_manager import (
     ClusterManager as CockroachDBClusterManager,
 )
@@ -67,6 +68,11 @@ def main():
             "params": [],
             "obj": ApiQuotaManager,
         },
+        "provision": {
+            "description": "Provision a BYODB database",
+            "params": [],
+            "obj": ProvisioningManager,
+        },
         "setup": {
             "description": "Execute suggested setup steps",
             "params": [],
@@ -96,6 +102,7 @@ def main():
         option = options[sys.argv[1]]
         params = option["params"]
         obj_cls = option["obj"]
+
         if len(params) > 0:
             if len(sys.argv) != 2 + len(params):
                 obj_cls(Config()).usage()